split up auth worker into files
